feat: return iterator at `vertices`, `adjacents` and `edges`
We define the `Item` of each iterator as `Cow<'_, ..>` because some
graph has adjacent of vertex directly, but some graph generate adjacent
from the vertex. It's good to return reference if we have the item, but
if we define `Item = &'_ ..`, this doesn't works if the item is created
inside of method because the reference of created item can't live
outside of the method. On the other hand, if we define `Item = ..`, it
will works on both cases, but cloning or copying the already exist item
is not efficient. So I define `Item = Cow<'_, ..>` so that user can
switch whether return reference of exist item or created item depende on
the implementation of graph.