feat: use `-> impl Iterator<..>` at some trait
This allow user to return iterator more easily. For example, the
following code
    fn vertices(&'a self) -> impl Iterator<Item = Cow<'a, V>> {
        self.vs.iter().map(Cow::Borrowed)
    }
is not easy at previous method. If we want to write such code, we need
write it inside Box or wrap iterator with struct and implement Iterator
to it to mapping emitted value. This need more amount of code to achieve
it. So, this change will good for competitive programming.